Cleanup and More Simplifications (#4)
